+++ CHANGELOG.md
@@ -7,6 +7,8 @@ x.y.z Release notes (yyyy-MM-dd)
 * When using `Realm.write(withoutNotifying:)` there was a chance that the 
   supplied observation blocks would not be skipped when in a write transaction.
   ([Object Store #1103](https://github.com/realm/realm-object-store/pull/1103))
+* Comparing two identical unmanaged `List<>`/`RLMArray` objects would fail.
+  ([#5665](https://github.com/realm/realm-cocoa/issues/5665)).
 
 <!-- ### Breaking Changes - ONLY INCLUDE FOR NEW MAJOR version -->